requirements.txt
azure-identity
azure-mgmt-network
azure-mgmt-compute
pytz
pyyaml